+++ libc/sysv/consts/falloc.h
@@ -4,13 +4,13 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long FALLOC_FL_KEEP_SIZE;
-extern const long FALLOC_FL_PUNCH_HOLE;
-extern const long FALLOC_FL_NO_HIDE_STALE;
-extern const long FALLOC_FL_COLLAPSE_RANGE;
-extern const long FALLOC_FL_ZERO_RANGE;
-extern const long FALLOC_FL_INSERT_RANGE;
-extern const long FALLOC_FL_UNSHARE_RANGE;
+extern const int FALLOC_FL_KEEP_SIZE;
+extern const int FALLOC_FL_PUNCH_HOLE;
+extern const int FALLOC_FL_NO_HIDE_STALE;
+extern const int FALLOC_FL_COLLAPSE_RANGE;
+extern const int FALLOC_FL_ZERO_RANGE;
+extern const int FALLOC_FL_INSERT_RANGE;
+extern const int FALLOC_FL_UNSHARE_RANGE;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
