+++ libc/sysv/consts/mfd.h
@@ -4,8 +4,8 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long MFD_CLOEXEC;
-extern const long MFD_ALLOW_SEALING;
+extern const unsigned int MFD_CLOEXEC;
+extern const unsigned int MFD_ALLOW_SEALING;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
