+++ libc/sysv/consts/scm.h
@@ -4,11 +4,11 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long SCM_TIMESTAMP;
-extern const long SCM_CREDENTIALS;
-extern const long SCM_TIMESTAMPING;
-extern const long SCM_TIMESTAMPNS;
-extern const long SCM_WIFI_STATUS;
+extern const int SCM_TIMESTAMP;
+extern const int SCM_CREDENTIALS;
+extern const int SCM_TIMESTAMPING;
+extern const int SCM_TIMESTAMPNS;
+extern const int SCM_WIFI_STATUS;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
