+++ libc/sysv/consts/shm.h
@@ -20,20 +20,20 @@
 #if !(__ASSEMBLER__ + __LINKER__ + 0)
 COSMOPOLITAN_C_START_
 
-extern const long SHM_DEST;
-extern const long SHM_EXEC;
-extern const long SHM_HUGETLB;
-extern const long SHM_INFO;
-extern const long SHM_LOCK;
-extern const long SHM_LOCKED;
-extern const long SHM_NORESERVE;
-extern const long SHM_R;
-extern const long SHM_RDONLY;
-extern const long SHM_REMAP;
-extern const long SHM_RND;
-extern const long SHM_STAT;
-extern const long SHM_UNLOCK;
-extern const long SHM_W;
+extern const int SHM_DEST;
+extern const int SHM_EXEC;
+extern const int SHM_HUGETLB;
+extern const int SHM_INFO;
+extern const int SHM_LOCK;
+extern const int SHM_LOCKED;
+extern const int SHM_NORESERVE;
+extern const int SHM_R;
+extern const int SHM_RDONLY;
+extern const int SHM_REMAP;
+extern const int SHM_RND;
+extern const int SHM_STAT;
+extern const int SHM_UNLOCK;
+extern const int SHM_W;
 
 COSMOPOLITAN_C_END_
 #endif /* !(__ASSEMBLER__ + __LINKER__ + 0) */
